fix: refactoring how the injectRefreshEntry func works
I didn't update the tests yet because I want to see if this is the right direction to go down before I finish the rest of it
