+++ lib/index.js
@@ -10,6 +10,7 @@ const {
   injectRefreshEntry,
   injectRefreshLoader,
   normalizeOptions,
+  getOverlayEntry,
 } = require('./utils');
 const schema = require('./options.json');
 
@@ -85,7 +86,10 @@ class ReactRefreshPlugin {
     }
 
     // Inject react-refresh context to all Webpack entry points
-    compiler.options.entry = injectRefreshEntry(compiler.options.entry, this.options, compiler);
+    compiler.options.entry = injectRefreshEntry(
+      compiler.options.entry,
+      ...getOverlayEntry(this.options, compiler.options.devServer || {})
+    );
 
     // Inject necessary modules to bundle's global scope
     /** @type {Record<string, string>} */
