+++ lib/utils/index.js
@@ -5,6 +5,7 @@ const getSocketIntegration = require('./getSocketIntegration');
 const injectRefreshEntry = require('./injectRefreshEntry');
 const injectRefreshLoader = require('./injectRefreshLoader');
 const normalizeOptions = require('./normalizeOptions');
+const getOverlayEntry = require('./getOverlayEntry');
 
 module.exports = {
   createError,
@@ -14,4 +15,5 @@ module.exports = {
   injectRefreshEntry,
   injectRefreshLoader,
   normalizeOptions,
+  getOverlayEntry,
 };
