+++ lib/utils/injectRefreshEntry.js
@@ -1,4 +1,3 @@
-const querystring = require('querystring');
 const createError = require('./createError');
 
 /** @typedef {string | string[] | import('webpack').Entry} StaticEntry */
@@ -28,55 +27,11 @@ function isSocketEntry(entry) {
 /**
  * Injects an entry to the bundle for react-refresh.
  * @param {WebpackEntry} [originalEntry] A Webpack entry object.
- * @param {import('../types').NormalizedPluginOptions} options Configuration options for this plugin.
- * @param {import('webpack').Compiler} compiler A webpack compiler object.
+ * @param {WebpackEntry} prependEntries A Webpack entry that will be prepended
+ * @param {WebpackEntry} overlayEntries A Webpack entry for the overlay
  * @returns {WebpackEntry} An injected entry object.
  */
-function injectRefreshEntry(originalEntry, options, compiler) {
-  /** @type {Record<string, *>} */
-  let resourceQuery = {};
-
-  if (compiler && compiler.options && compiler.options.devServer) {
-    const { sockHost, sockPath, sockPort } = compiler.options.devServer;
-
-    sockHost && (resourceQuery.sockHost = sockHost);
-    sockPath && (resourceQuery.sockPath = sockPath);
-    sockPort && (resourceQuery.sockPort = sockPort);
-  }
-
-  if (options.overlay) {
-    options.overlay.sockHost && (resourceQuery.sockHost = options.overlay.sockHost);
-    options.overlay.sockPath && (resourceQuery.sockPath = options.overlay.sockPath);
-    options.overlay.sockPort && (resourceQuery.sockPort = options.overlay.sockPort);
-  }
-
-  // We don't need to URI encode the resourceQuery as it will be parsed by Webpack
-  const queryString = querystring.stringify(resourceQuery, undefined, undefined, {
-    /**
-     * @param {string} string
-     * @returns {string}
-     */
-    encodeURIComponent(string) {
-      return string;
-    },
-  });
-
-  const prependEntries = [
-    // React-refresh runtime
-    require.resolve('../../client/ReactRefreshEntry'),
-  ];
-
-  const overlayEntries = [
-    // Legacy WDS SockJS integration
-    options.overlay &&
-      options.overlay.useLegacyWDSSockets &&
-      require.resolve('../../client/LegacyWDSSocketEntry'),
-    // Error overlay runtime
-    options.overlay &&
-      options.overlay.entry &&
-      options.overlay.entry + (queryString && `?${queryString}`),
-  ].filter(Boolean);
-
+function injectRefreshEntry(originalEntry, prependEntries, overlayEntries) {
   // Single string entry point
   if (typeof originalEntry === 'string') {
     if (isSocketEntry(originalEntry)) {
@@ -105,9 +60,9 @@ function injectRefreshEntry(originalEntry, options, compiler) {
           typeof curEntry === 'object' && curEntry.import
             ? {
                 ...curEntry,
-                import: injectRefreshEntry(curEntry.import, options),
+                import: injectRefreshEntry(curEntry.import, prependEntries, overlayEntries),
               }
-            : injectRefreshEntry(curEntry, options),
+            : injectRefreshEntry(curEntry, prependEntries, overlayEntries),
       }),
       {}
     );
@@ -116,7 +71,7 @@ function injectRefreshEntry(originalEntry, options, compiler) {
   if (typeof originalEntry === 'function') {
     return (...args) =>
       Promise.resolve(originalEntry(...args)).then((resolvedEntry) =>
-        injectRefreshEntry(resolvedEntry, options)
+        injectRefreshEntry(resolvedEntry, prependEntries, overlayEntries)
       );
   }
